requirements.txt
langchain
langchain-core
langchain-community
langchain-text-splitters
pypdf
pymupdf
sentence-transformers
faiss-cpu
chromadb
ipykernel
langchain-google-genai
langchain-deepseek
langchain-openai
python-dotenv